Create oj_pancakes
describe
Today, TheBeet went to the shandong pancake stall at the gate to buy pancakes. When he paid for the pancake, he found that there was only a piece of 100 yuan in his wallet.So he needed to break the $100 bill, but TheBeet didn't like COINS, and for some reason TheBeet didn't like the $20 bill, and he didn't like the fact that there were too many notes in the wallet, so he wanted as few notes as possible.Give the number of notes the owner has and ask the owner how to give TheBeet change to meet his demand.
The input
The first line of input is a positive integer N(1 <= N <= 40), indicating that TheBeet has bought N shandong pancakes, and it is known that each shandong pancake costs 2.5 yuan.Then there are 6 non-negative integers within the range of [0, 1000], which respectively represent the number of 50 yuan, 10 yuan, 5 yuan, 1 yuan, 5 jiao and 1 jiao notes owned by the stall owner.
The output
Print an integer representing the number of notes TheBeet received last.If it doesn't, it prints -1.