Create oj_longest
describe
Legend has it that matchmaker is the god of marriage.Every year on the Chinese valentine's day, the seven-star lady will make a list of unmarried adult men and women in the world and report it to the heaven.After the matchmaker receives the roster, according to the personality, the good and evil, the interest and the condition copies writes a spouse roster, then ties the male and female two people's feet with the red line, makes the suitable male and female matches a good match.In an old town, an old river runs north and south, dividing the town into east and west parts.The old town kept an old custom. All the unmarried men lived on the west bank of the river, and all the unmarried women lived on the east bank.This year, he received A list of all the unmarried men and women in the town. He made A simple summary of each person's personality, good and evil, interests and conditions, and added A tag like A, B and C to each person.Originally this is a very simple thing, but now the problem is the red line used by the matchmaker, can not cross each other, otherwise the consequences will be very serious.So matchmaker in order to more people's happiness, he can only sacrifice part of the people's happiness.Now matchmaker on the paper for a long time, or can not figure out a best plan, so that the most couples get married.Can you help him?
The input
The first line of the input file contains two integers, N and M(0 < N,M <= 1000), representing the number of unmarried men and unmarried women, respectively.The second line contains A string of length N, consisting of the uppercase a-z and lowercase a-z, with the north to south order representing each man's tag.The third line contains a string of length M that represents each woman's tag in north to south order.
The output
The output of an integer indicates the maximum number of couples a matchmaker can successfully match.